# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Datasur SMS API",
        "version": "0.6 alpha"
    },
    "paths": {
        "/api": {},
        "/api/login": {
            "post": {
                "summary": "Attempt user login by providing email and password. Bearer token is returned.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "johndoe@contoso.com",
                                    "password": "contoso123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "summary": "View details of current signed in user.",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/history": {
            "get": {
                "summary": "View send-history of signed-in user",
                "operationId": "2c91f335de85204f3ae1465ff34e29d2",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/history/{smsUuid}": {
            "get": {
                "summary": "View details of an SMS by providing Uuid",
                "operationId": "0ad70b8fc07f0c7f97ee2f4632f165de",
                "parameters": [
                    {
                        "name": "smsUuid",
                        "in": "path",
                        "description": "SMS Uuid parameter",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "An SMS UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/history/{smsUuid}/isSent": {
            "get": {
                "summary": "Validate if an SMS has been sent by sending the SMS Uuid",
                "operationId": "e80d1427c0c940f7df6e735f0eccbf42",
                "parameters": [
                    {
                        "name": "smsUuid",
                        "in": "path",
                        "description": "SMS Uuid parameter",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "An SMS UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/sms/send/now": {
            "post": {
                "summary": "Sends SMS now",
                "operationId": "9ff0b646c6ca062894a73f23c8ff8528",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "recipient": {
                                        "type": "integer"
                                    },
                                    "body": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "recipient": 5977654321,
                                    "body": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/sms/send/later": {
            "post": {
                "summary": "Sends SMS later",
                "operationId": "296a92a6e5262443ea76cf05d55204f9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "recipient": {
                                        "type": "integer"
                                    },
                                    "body": {
                                        "type": "string"
                                    },
                                    "datetime": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "recipient": 5977654321,
                                    "body": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do",
                                    "datetime": "1975-12-25 14:15:16"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/sms/cancel/{smsUuid}": {
            "post": {
                "summary": "Cancel an SMS by providing SMS Uuid",
                "operationId": "70537cac5b6b0bd8a826f349c275652d",
                "parameters": [
                    {
                        "name": "smsUuid",
                        "in": "path",
                        "description": "SMS Uuid parameter",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "An SMS UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/sms/bulksend/now": {
            "post": {
                "summary": "Sends SMS to multiple receipients now",
                "operationId": "bf13256931dcd25756dacdebffc7990a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "recipients": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "body": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "recipients": [
                                        5977654321,
                                        5971234567,
                                        5978765432
                                    ],
                                    "body": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/sms/bulksend/later": {
            "post": {
                "summary": "Sends SMS to multiple receipients later",
                "operationId": "dcece7b694138057d4c68ecd0be3216f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "recipients": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "body": {
                                        "type": "string"
                                    },
                                    "datetime": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "recipients": [
                                        5977654321,
                                        5971234567,
                                        5978765432
                                    ],
                                    "body": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do",
                                    "datetime": "1975-12-25 14:15:16"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/bulksend/cancel/{smsUuid}": {
            "post": {
                "summary": "Cancel a bulk send by providing Uuid of assignment",
                "operationId": "2b8767c0abae04d314e1da9c5c0c5369",
                "parameters": [
                    {
                        "name": "bulkUuid",
                        "in": "path",
                        "description": "Bulk Uuid parameter",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "A bulk UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}